pack/main: Silence MSVC compiler warning

+++ src/pack/main.cpp
@@ -70,7 +70,7 @@ addFile(CreateArchiveContext& ctx, StringView path) noexcept {
     uiShowAddedFile(path, data_.size());
 
     LockGuard guard(ctx.packMutex);
-    ctx.pack->addBlob(move_(path), data_.size(), data_.data());
+    ctx.pack->addBlob(move_(path), static_cast<uint32_t>(data_.size()), data_.data());
 
     data_.reset_lose_memory();  // Don't delete data pointer.
 }
